2338. Sub 4

+++ rust/src/_2338_count_the_number_of_ideal_arrays.rs
@@ -40,13 +40,13 @@ impl Solution {
 
         let mut result = 0;
         let mut combo = 1;
-        let mut top = n as i64 - 1;
+        let mut top = n as i128 - 1;
         let mut bottom = 1;
         let mut base = 1;
         for strict_count in strict_counts.iter().cloned() {
             if base <= max_value {
-                result = ((result + combo * strict_count[(max_value - base) as usize] as i64)
-                    % MODULO as i64) as _;
+                result = ((result + combo * strict_count[(max_value - base) as usize] as i128)
+                    % MODULO as i128) as _;
             } else {
                 break;
             }
@@ -74,6 +74,7 @@ mod tests {
     #[case(5, 3, 11)]
     #[case(5, 9, 111)]
     #[case(184, 389, 510488787)]
+    #[case(5878, 2900, 465040898)]
     fn case(#[case] n: i32, #[case] max_value: i32, #[case] expected: i32) {
         let actual = Solution::ideal_arrays(n, max_value);
         assert_eq!(actual, expected);
